+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 * Added option "Both direction" into "Expand to".
+* Update button grab system.
+* Added buttons "MinimapZoomIn", "MinimapZoomOut", "MiniMapWorldMapButton" for grab using "Grab default buttons".